Restructured the Visual components into separate classes
PeptideSpectrumMatchPlot only draws the spectrum shown
CrosslinkSpectrumMatch only adds onto the PepetideSpectrumMatchPlot
DrawnSequences takes over the above twos functionality of drawing sequences
Adjusted some tests to account for these changes in structure
