+++ Test/MetaDrawTest.cs
@@ -1,6 +1,7 @@
 ﻿using Chemistry;
 using EngineLayer;
 using GuiFunctions;
+using GuiFunctions.MetaDraw;
 using MassSpectrometry;
 using NUnit.Framework;
 using OxyPlot.Series;
@@ -136,7 +137,7 @@ namespace Test
             var parentChildView = new ParentChildScanPlotsView();
             MetaDrawSettings.FirstAAonScreenIndex = 0;
             MetaDrawSettings.NumberOfAAOnScreen = 20; // Will be dynamic based upon window size, 20 is an arbitraty number used for testing purposes
-
+            MetaDrawSettings.DrawMatchedIons = true;
             int numAnnotatedResidues = psm.BaseSeq.Length;
             int numAnnotatedIons = psm.MatchedIons.Count;
             int numAnnotatedMods = psm.FullSequence.Count(p => p == '[');
@@ -145,15 +146,16 @@ namespace Test
             // Iterates through the psm, simulating scrolling, until the sequence is scrolled as far as allowed
             for (; MetaDrawSettings.FirstAAonScreenIndex < psm.BaseSeq.Length - MetaDrawSettings.NumberOfAAOnScreen; MetaDrawSettings.FirstAAonScreenIndex++)
             {
-                metadrawLogic.DisplaySpectrumMatch(plotView, stationaryCanvas, psm, parentChildView, out errors, scrollableCanvas);
+                metadrawLogic.DisplaySequences(stationaryCanvas, scrollableCanvas, psm);
+                metadrawLogic.DisplaySpectrumMatch(plotView, psm, parentChildView, out errors);
                 Assert.That(errors == null || !errors.Any());
                 // Checks to see if scrollable sequence is the same each time
-                Assert.That(scrollableCanvas.Children.Count == numAnnotatedResidues + numAnnotatedIons + numAnnotatedMods);
+                Assert.That(metadrawLogic.ScrollableSequence.SequenceDrawingCanvas.Children.Count == numAnnotatedResidues + numAnnotatedIons + numAnnotatedMods);
 
                 // Checks to see if the stationary sequence updated with the new positioning
                 string modifiedBaseSeq = psm.BaseSeq.Substring(MetaDrawSettings.FirstAAonScreenIndex, MetaDrawSettings.NumberOfAAOnScreen);
                 string fullSequence = psm.BaseSeq;
-                Dictionary<int, string> modDictionary = PeptideSpectrumMatchPlot.ParseModifications(psm.FullSequence);
+                Dictionary<int, string> modDictionary = DrawnSequence.ParseModifications(psm.FullSequence);
                 foreach (var mod in modDictionary)
                 {
                     if (mod.Key >= MetaDrawSettings.FirstAAonScreenIndex && mod.Key < (MetaDrawSettings.FirstAAonScreenIndex + MetaDrawSettings.NumberOfAAOnScreen))
@@ -163,7 +165,7 @@ namespace Test
                 }
                 List<MatchedFragmentIon> matchedIons = psm.MatchedIons.Where(p => p.NeutralTheoreticalProduct.AminoAcidPosition > MetaDrawSettings.FirstAAonScreenIndex &&
                                                        p.NeutralTheoreticalProduct.AminoAcidPosition < (MetaDrawSettings.FirstAAonScreenIndex + MetaDrawSettings.NumberOfAAOnScreen)).ToList();
-                Assert.That(stationaryCanvas.Children.Count == modifiedBaseSeq.Length + matchedIons.Count + fullSequence.Count(p => p == '['));
+                Assert.That(metadrawLogic.StationarySequence.SequenceDrawingCanvas.Children.Count == modifiedBaseSeq.Length + matchedIons.Count + fullSequence.Count(p => p == '['));
             }
         }
 
@@ -239,7 +241,9 @@ namespace Test
 
             MetaDrawSettings.FirstAAonScreenIndex = 0;
             MetaDrawSettings.NumberOfAAOnScreen = psm.BaseSeq.Length;
-            metadrawLogic.DisplaySpectrumMatch(plotView, stationaryCanvas, psm, parentChildView, out errors, scrollableCanvas);
+            MetaDrawSettings.DrawMatchedIons = true;
+            metadrawLogic.DisplaySequences(stationaryCanvas, scrollableCanvas, psm);
+            metadrawLogic.DisplaySpectrumMatch(plotView, psm, parentChildView, out errors);
             Assert.That(errors == null || !errors.Any());
 
             // test that plot was drawn
@@ -258,12 +262,12 @@ namespace Test
             int numAnnotatedResidues = psm.BaseSeq.Length;
             int numAnnotatedIons = psm.MatchedIons.Count;
             int numAnnotatedMods = psm.FullSequence.Count(p => p == '[');
-            Assert.That(scrollableCanvas.Children.Count == numAnnotatedResidues + numAnnotatedIons + numAnnotatedMods);
+            Assert.That(metadrawLogic.ScrollableSequence.SequenceDrawingCanvas.Children.Count == numAnnotatedResidues + numAnnotatedIons + numAnnotatedMods);
 
             // test that the stationary sequence annotation was drawn
             string modifiedBaseSeq = psm.BaseSeq.Substring(MetaDrawSettings.FirstAAonScreenIndex, MetaDrawSettings.NumberOfAAOnScreen);
             string fullSequence = psm.BaseSeq;
-            Dictionary<int, string> modDictionary = PeptideSpectrumMatchPlot.ParseModifications(psm.FullSequence);
+            Dictionary<int, string> modDictionary = DrawnSequence.ParseModifications(psm.FullSequence);
             foreach (var mod in modDictionary)
             {
                 if (mod.Key >= MetaDrawSettings.FirstAAonScreenIndex && mod.Key < (MetaDrawSettings.FirstAAonScreenIndex + MetaDrawSettings.NumberOfAAOnScreen))
@@ -273,7 +277,7 @@ namespace Test
             }
             List<MatchedFragmentIon> matchedIons = psm.MatchedIons.Where(p => p.NeutralTheoreticalProduct.AminoAcidPosition > MetaDrawSettings.FirstAAonScreenIndex &&
                                                    p.NeutralTheoreticalProduct.AminoAcidPosition < (MetaDrawSettings.FirstAAonScreenIndex + MetaDrawSettings.NumberOfAAOnScreen)).ToList();
-            Assert.That(stationaryCanvas.Children.Count == modifiedBaseSeq.Length + matchedIons.Count + fullSequence.Count(p => p == '[' ));
+            Assert.That(metadrawLogic.StationarySequence.SequenceDrawingCanvas.Children.Count == modifiedBaseSeq.Length + matchedIons.Count + fullSequence.Count(p => p == '[' ));
 
             // write pdf
             var psmsToExport = metadrawLogic.FilteredListOfPsms.Where(p => p.FullSequence == "QIVHDSGR").Take(3).ToList();
@@ -288,12 +292,13 @@ namespace Test
             // test displaying a PSM with a mod
             var modPsm = metadrawLogic.FilteredListOfPsms.First(p => p.FullSequence.Contains("["));
             MetaDrawSettings.NumberOfAAOnScreen = modPsm.BaseSeq.Length;
-            metadrawLogic.DisplaySpectrumMatch(plotView, stationaryCanvas, modPsm, parentChildView, out errors, scrollableCanvas);
+            metadrawLogic.DisplaySequences(stationaryCanvas, scrollableCanvas, modPsm);
+            metadrawLogic.DisplaySpectrumMatch(plotView, modPsm, parentChildView, out errors);
             Assert.That(errors == null || !errors.Any());
-            Assert.That(scrollableCanvas.Children.Count == modPsm.BaseSeq.Length + modPsm.MatchedIons.Count + modPsm.FullSequence.Count(p => p == '['));
+            Assert.That(metadrawLogic.ScrollableSequence.SequenceDrawingCanvas.Children.Count == modPsm.BaseSeq.Length + modPsm.MatchedIons.Count + modPsm.FullSequence.Count(p => p == '['));
             modifiedBaseSeq = modPsm.BaseSeq.Substring(MetaDrawSettings.FirstAAonScreenIndex, MetaDrawSettings.NumberOfAAOnScreen);
             fullSequence = modPsm.BaseSeq;
-            modDictionary = PeptideSpectrumMatchPlot.ParseModifications(modPsm.FullSequence);
+            modDictionary = DrawnSequence.ParseModifications(modPsm.FullSequence);
             foreach (var mod in modDictionary)
             {
                 if (mod.Key >= MetaDrawSettings.FirstAAonScreenIndex && mod.Key < (MetaDrawSettings.FirstAAonScreenIndex + MetaDrawSettings.NumberOfAAOnScreen))
@@ -303,7 +308,7 @@ namespace Test
             }
             matchedIons = modPsm.MatchedIons.Where(p => p.NeutralTheoreticalProduct.AminoAcidPosition > MetaDrawSettings.FirstAAonScreenIndex &&
                                                 p.NeutralTheoreticalProduct.AminoAcidPosition < (MetaDrawSettings.FirstAAonScreenIndex + MetaDrawSettings.NumberOfAAOnScreen)).ToList();
-            Assert.That(stationaryCanvas.Children.Count == modifiedBaseSeq.Length + matchedIons.Count + fullSequence.Count(p => p == '['));
+            Assert.That(metadrawLogic.StationarySequence.SequenceDrawingCanvas.Children.Count == modifiedBaseSeq.Length + matchedIons.Count + fullSequence.Count(p => p == '['));
 
             // clean up resources
             metadrawLogic.CleanUpResources();
@@ -377,12 +382,13 @@ namespace Test
 
             // draw PSM
             var plotView = new OxyPlot.Wpf.PlotView();
-            var canvas = new Canvas();
+            var stationaryCanvas = new Canvas();
             var scrollableCanvas = new Canvas();
             var parentChildView = new ParentChildScanPlotsView();
             var csm = metadrawLogic.FilteredListOfPsms.First();
 
-            metadrawLogic.DisplaySpectrumMatch(plotView, canvas, csm, parentChildView, out errors, scrollableCanvas);
+            metadrawLogic.DisplaySequences(stationaryCanvas, scrollableCanvas, csm);
+            metadrawLogic.DisplaySpectrumMatch(plotView, csm, parentChildView, out errors);
             Assert.That(errors == null || !errors.Any());
 
             // test that plot was drawn
@@ -397,7 +403,7 @@ namespace Test
             Assert.That(plotAxes.Count == 2);
 
             // test that base sequence annotation was drawn
-            Assert.That(canvas.Children.Count > 0);
+            Assert.That(stationaryCanvas.Children.Count > 0);
 
             // test that the plots were drawn in the parent/child view
             Assert.That(parentChildView.Plots.Count == 2);
@@ -436,7 +442,7 @@ namespace Test
 
             // write pdf
             var psmsToExport = metadrawLogic.FilteredListOfPsms.Where(p => p.FullSequence == "SLGKVGTR(4)").ToList();
-            metadrawLogic.ExportToPdf(plotView, canvas, psmsToExport, parentChildView, outputFolder, out errors);
+            metadrawLogic.ExportToPdf(plotView, metadrawLogic.StationarySequence.SequenceDrawingCanvas, psmsToExport, parentChildView, outputFolder, out errors);
 
             // test that pdf exists
             Assert.That(File.Exists(Path.Combine(outputFolder, @"2_SLGKVGTR(4).pdf"))); // parent scan
@@ -512,13 +518,14 @@ namespace Test
 
             // draw PSM
             var plotView = new OxyPlot.Wpf.PlotView();
+            var parentChildView = new ParentChildScanPlotsView();
             var stationaryCanvas = new Canvas();
             var scrollableCanvas = new Canvas();
-            var parentChildView = new ParentChildScanPlotsView();
             var psm = metadrawLogic.FilteredListOfPsms.First();
             MetaDrawSettings.NumberOfAAOnScreen = psm.BaseSeq.Length;
 
-            metadrawLogic.DisplaySpectrumMatch(plotView, stationaryCanvas, psm, parentChildView, out errors, scrollableCanvas);
+            metadrawLogic.DisplaySequences(stationaryCanvas, scrollableCanvas, psm);
+            metadrawLogic.DisplaySpectrumMatch(plotView, psm, parentChildView, out errors);
             Assert.That(errors == null || !errors.Any());
 
             // test that plot was drawn
@@ -533,7 +540,7 @@ namespace Test
             Assert.AreEqual(2, plotAxes.Count);
 
             // test that base sequence annotation was drawn
-            Assert.Greater(stationaryCanvas.Children.Count, 0);
+            Assert.Greater(metadrawLogic.StationarySequence.SequenceDrawingCanvas.Children.Count, 0);
 
             // test that the plots were drawn in the parent/child view
             Assert.AreEqual(2, parentChildView.Plots.Count);
@@ -575,7 +582,7 @@ namespace Test
 
             // write pdf
             var psmsToExport = metadrawLogic.FilteredListOfPsms.Where(p => p.FullSequence == "STTAVQTPTSGEPLVST[O-Glycosylation:H1N1 on X]SEPLSSK").ToList();
-            metadrawLogic.ExportToPdf(plotView, stationaryCanvas, psmsToExport, parentChildView, outputFolder, out errors, scrollableCanvas);
+            metadrawLogic.ExportToPdf(plotView, metadrawLogic.StationarySequence.SequenceDrawingCanvas, psmsToExport, parentChildView, outputFolder, out errors);
 
             // test that pdf exists
             Assert.That(File.Exists(Path.Combine(outputFolder, @"27_STTAVQTPTSGEPLVST[O-Glycosylat.pdf"))); // parent scan
@@ -626,7 +633,7 @@ namespace Test
             var parentChildScanPlotsView = new ParentChildScanPlotsView();
 
             // plotting PSM should produce an error because spectra are not loaded
-            metadrawLogic.DisplaySpectrumMatch(plotView, canvas, psmsFromTsv.First(), parentChildScanPlotsView, out errors, scrollableCanvas);
+            metadrawLogic.DisplaySpectrumMatch(plotView, psmsFromTsv.First(), parentChildScanPlotsView, out errors);
             Assert.That(errors.Any());
 
             // export to PDF should produce an error because spectra are not loaded
@@ -672,11 +679,13 @@ namespace Test
             var plotView = new OxyPlot.Wpf.PlotView();
             var canvas = new Canvas();
             var scrollableCanvas = new Canvas();
+            var stationaryCanvas = new Canvas();
             var parentChildScanPlotsView = new ParentChildScanPlotsView();
 
             // plot PSM
             MetaDrawSettings.NumberOfAAOnScreen = metadrawLogic.FilteredListOfPsms.First().BaseSeq.Length;
-            metadrawLogic.DisplaySpectrumMatch(plotView, canvas, metadrawLogic.FilteredListOfPsms.First(), parentChildScanPlotsView, out errors, scrollableCanvas);
+            metadrawLogic.DisplaySequences(stationaryCanvas, scrollableCanvas, metadrawLogic.FilteredListOfPsms.First());
+            metadrawLogic.DisplaySpectrumMatch(plotView, metadrawLogic.FilteredListOfPsms.First(), parentChildScanPlotsView, out errors);
             Assert.That(errors == null || !errors.Any());
 
             // export to PDF
@@ -729,11 +738,13 @@ namespace Test
             var plotView = new OxyPlot.Wpf.PlotView();
             var canvas = new Canvas();
             var scrollableCanvas = new Canvas();
+            var stationaryCanvas = new Canvas();
             var parentChildView = new ParentChildScanPlotsView();
             var psm = metadrawLogic.FilteredListOfPsms.First();
 
             MetaDrawSettings.NumberOfAAOnScreen = metadrawLogic.FilteredListOfPsms.First().BaseSeq.Length;
-            metadrawLogic.DisplaySpectrumMatch(plotView, canvas, psm, parentChildView, out errors, scrollableCanvas);
+            metadrawLogic.DisplaySequences(stationaryCanvas, scrollableCanvas, psm);
+            metadrawLogic.DisplaySpectrumMatch(plotView, psm, parentChildView, out errors);
             Assert.That(errors == null || !errors.Any());
 
             // test that plot was drawn
